\name{example_annotation}
\docType{data}
\alias{example_annotation}
\title{Example Annotation Data for PersomicsArray Package}
\description{
  This is a very small example of the expected array annotation format
  taken as input by the spot_id function in the PersomicsArray pacakge.
}
\usage{example_annotation}
\format{data.frame}
\keyword{datasets}
\seealso{
<\code{\link{PersomicsArray-package}}>
<\code{\link{spot_id}}>
<\code{\link{plot_low_res}}>
<\code{\link{example_plate}}>
}
