\name{example_plate}
\docType{data}
\alias{example_plate}
\title{Example Plate Image for PersomicsArray Package}
\description{
  This is a very small example image used for teaching functionality
  in the PersomicsArray pacakge.
}
\usage{example_plate}
\format{data.frame}
\keyword{datasets}
\seealso{
<\code{\link{PersomicsArray-package}}>
<\code{\link{spot_id}}>
<\code{\link{plot_low_res}}>
<\code{\link{example_annotation}}>
}
